+++ tests/testenv/entrypoint.sh
@@ -32,6 +32,7 @@ setLDAP () {
     # Create base entry and populate the dictionary.
     ldapadd -x -D "cn=admin,dc=bonsai,dc=test" -w p@ssword -H ldapi:/// -f /root/base.ldif
     ldapadd -x -D "cn=admin,dc=bonsai,dc=test" -w p@ssword -H ldapi:/// -f /root/users.ldif
+    ldapadd -x -D "cn=admin,dc=bonsai,dc=test" -w p@ssword -H ldapi:/// -f /root/referrals.ldif
     # Set default password policy.
     ldapadd -x -D "cn=admin,dc=bonsai,dc=test" -w p@ssword -H ldapi:/// -f /root/ppolicy.ldif
     # Stop the slapd.